requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
iniconfig==1.1.1
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
tomli==2.0.1
tweepy==4.7.0
urllib3==1.26.9
